Lima: Use data volume for K3s data
This sets up a provision script which sets up bind mounts on boot so that
/etc/rancher and /var/lib/rancher are forwarded to the /mnt/data volume, so
that they can be persisted as well as have more room for contents.  To make
the script somewhat managable, the default Lima configuration is now stored
in a YAML file rather than inline within the source.  We still need to
override config options at runtime though.

Signed-off-by: Mark Yen <mark.yen@suse.com>