Do not check in `go.work.sum`
That file is only used for modules that are _not_ listed in `go.sum` files
of the individual modules.  In practice, now that we auto-sync `go.mod`
across PRs, that is only used when the developer's IDE installs things that
are not used in the actual project.  Exclude that file from git (but keep
committing `go.work` to ensure things stay in sync).

Signed-off-by: Mark Yen <mark.yen@suse.com>