BATS: Remove use of `start_application`
`start_application` was ambiguous, and immediately waited for kubelet; that
often means we run too slow in CI and time out.  Instead, have the callers
call `start_container_engine` or `start_kubernetes` as appropriate (since
not starting Kubernetes makes the tests run faster), and wait for the
container engine before waiting for kubelet to ensure we record forward
progress.

Signed-off-by: Mark Yen <mark.yen@suse.com>