Replace yarn lint:spelling with a shell script
This is useful because we're removing the action (and running yarn in CI
instead); so the configuration we had before is going away.

Currently, running this on Windows is not supported; however, we may be
able to end up doing it via WSL2 at some point.  (Currently there seems to
be issues with Perl.)

Signed-off-by: Mark Yen <mark.yen@suse.com>