Scripts: Only build included files when building main/preload
Previously we were pointlessly compiling everything (including the
frontend) when building the main and preload scripts.  This causes issues
when working on the frontend because we didn't actually have the vue loader
active.

This also fixes an issue in the preload script building where we never
found the ts-loader configuration (and therefore didn't modify it).

Signed-off-by: Mark Yen <mark.yen@suse.com>