Screenshots: Improve reliability
It seems like we some times end up moving to the volumes tab in the virtual
machine preferences, and _then_ check that the VZ radio in the emulation
tab was clicked.  Try to fix things by moving the clicking of the VZ radio
into the previous step (on the emulation tab), in the hopes that Playwright
will actually finish that before proceeding with the next test.

Signed-off-by: Mark Yen <mark.yen@suse.com>