CI: go-work-sync: Avoid post-install scripts
Changes to the `go.mod` files may update `go` lines (to require a different
compiler); avoid running the post-install scripts (which we do not require
in this workflow) so we can actually fix the code.

Also update `yarn lint:go:fix` to run `go work use` to update `go.work`
before we run `go list -m ...` which would error out if that update was
needed.

Signed-off-by: Mark Yen <mark.yen@suse.com>